remove ExpressionUtils and move it to be public apis in ChainStringBuilderAppendCalls
