+++ rewrite-java/src/main/java/org/openrewrite/java/tree/Space.java
@@ -34,7 +34,6 @@ import static java.util.Collections.emptyList;
 @EqualsAndHashCode
 public class Space {
     public static final Space EMPTY = new Space("", emptyList());
-    public static final Space SINGLE_SPACE = new Space(" ", emptyList());
 
     private final List<Comment> comments;
